Version 2.0.0 - Alpha 5

+++ Bybit.Api/Models/Market/BybitRiskLimit.cs
@@ -1,13 +1,43 @@
 ﻿namespace Bybit.Api.Models.Market;
 
+/// <summary>
+/// Bybit Risk Limit
+/// </summary>
 public class BybitRiskLimit
 {
-    public long Id { get; set; }
+    /// <summary>
+    /// Risk limit id
+    /// </summary>
+    public int Id { get; set; }
+
+    /// <summary>
+    /// Symbol
+    /// </summary>
     public string Symbol { get; set; }
+
+    /// <summary>
+    /// Risk limit value
+    /// </summary>
     public decimal RiskLimitValue { get; set; }
+
+    /// <summary>
+    /// Maintenance margin
+    /// </summary>
     public decimal MaintenanceMargin { get; set; }
+
+    /// <summary>
+    /// Initial margin
+    /// </summary>
     public decimal InitialMargin { get; set; }
+
+    /// <summary>
+    /// Is lowest risk
+    /// </summary>
     [JsonConverter(typeof(BooleanConverter))]
     public bool IsLowestRisk { get; set; }
+
+    /// <summary>
+    /// Max leverage
+    /// </summary>
     public decimal MaxLeverage { get; set; }
 }
